app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
PyJWT
passlib
pytest
httpcore
httpx
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
twisted
service_identity
requests
cryptography